pyproject.toml
[project]
name = "gcp-risk-advisory-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary>=2.9.12",
    "requests>=2.33.1",
]
